+++ lib/Doctrine/DBAL/Platforms/SQLServerPlatform.php
@@ -14,7 +14,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * This software consists of voluntary contributions made by many individuals
- * and is licensed under the LGPL. For more information, see
+ * and is licensed under the MIT license. For more information, see
  * <http://www.doctrine-project.org>.
  */
 
@@ -26,17 +26,43 @@ use Doctrine\DBAL\Schema\Index,
     Doctrine\DBAL\Schema\Table;
 
 /**
- * The MsSqlPlatform provides the behavior, features and SQL dialect of the
- * MySQL database platform.
+ * The SQLServerPlatform provides the behavior, features and SQL dialect of the
+ * Microsoft SQL Server database platform.
  *
  * @since 2.0
  * @author Roman Borschel <roman@code-factory.org>
  * @author Jonathan H. Wage <jonwage@gmail.com>
  * @author Benjamin Eberlei <kontakt@beberlei.de>
- * @todo Rename: MsSQLPlatform
  */
-class MsSqlPlatform extends AbstractPlatform
+class SQLServerPlatform extends AbstractPlatform
 {
+    /**
+     * {@inheritDoc}
+     */
+    public function getDateDiffExpression($date1, $date2)
+    {
+        return 'DATEDIFF(day, ' . $date2 . ',' . $date1 . ')';
+    }
+
+    public function getDateAddDaysExpression($date, $days)
+    {
+        return 'DATEADD(day, ' . $days . ', ' . $date . ')';
+    }
+
+    public function getDateSubDaysExpression($date, $days)
+    {
+        return 'DATEADD(day, -1 * ' . $days . ', ' . $date . ')';
+    }
+
+    public function getDateAddMonthExpression($date, $months)
+    {
+        return 'DATEADD(month, ' . $months . ', ' . $date . ')';
+    }
+
+    public function getDateSubMonthExpression($date, $months)
+    {
+        return 'DATEADD(month, -1 * ' . $months . ', ' . $date . ')';
+    }
 
     /**
      * Whether the platform prefers identity columns for ID generation.
@@ -127,7 +153,6 @@ class MsSqlPlatform extends AbstractPlatform
     public function getDropIndexSQL($index, $table=null)
     {
         if ($index instanceof \Doctrine\DBAL\Schema\Index) {
-            $index_ = $index;
             $index = $index->getQuotedName($this);
         } else if (!is_string($index)) {
             throw new \InvalidArgumentException('AbstractPlatform::getDropIndexSQL() expects $index parameter to be string or \Doctrine\DBAL\Schema\Index.');
@@ -169,7 +194,11 @@ class MsSqlPlatform extends AbstractPlatform
         }
 
         if (isset($options['primary']) && !empty($options['primary'])) {
-            $columnListSql .= ', PRIMARY KEY(' . implode(', ', array_unique(array_values($options['primary']))) . ')';
+            $flags = '';
+            if (isset($options['primary_index']) && $options['primary_index']->hasFlag('nonclustered')) {
+                $flags = ' NONCLUSTERED';
+            }
+            $columnListSql .= ', PRIMARY KEY' . $flags . ' (' . implode(', ', array_unique(array_values($options['primary']))) . ')';
         }
 
         $query = 'CREATE TABLE ' . $tableName . ' (' . $columnListSql;
@@ -183,18 +212,34 @@ class MsSqlPlatform extends AbstractPlatform
         $sql[] = $query;
 
         if (isset($options['indexes']) && !empty($options['indexes'])) {
-            foreach ($options['indexes'] AS $index) {
+            foreach ($options['indexes'] as $index) {
                 $sql[] = $this->getCreateIndexSQL($index, $tableName);
             }
         }
 
         if (isset($options['foreignKeys'])) {
-            foreach ((array) $options['foreignKeys'] AS $definition) {
+            foreach ((array) $options['foreignKeys'] as $definition) {
                 $sql[] = $this->getCreateForeignKeySQL($definition, $tableName);
             }
         }
 
         return $sql;
+    }
+
+    /**
+     * Get SQL to create an unnamed primary key constraint.
+     *
+     * @param Index $index
+     * @param string|Table $table
+     * @return string
+     */
+    public function getCreatePrimaryKeySQL(Index $index, $table)
+    {
+        $flags = '';
+        if ($index->hasFlag('nonclustered')) {
+            $flags = ' NONCLUSTERED';
+        }
+        return 'ALTER TABLE ' . $table . ' ADD PRIMARY KEY' . $flags . ' (' . $this->getIndexFieldDeclarationListSQL($index->getColumns()) . ')';
     }
 
     /**
@@ -224,6 +269,25 @@ class MsSqlPlatform extends AbstractPlatform
     }
 
     /**
+     * @override
+     */
+    protected function getCreateIndexSQLFlags(Index $index)
+    {
+        $type = '';
+        if ($index->isUnique()) {
+            $type .= 'UNIQUE ';
+        }
+
+        if ($index->hasFlag('clustered')) {
+            $type .= 'CLUSTERED ';
+        } else if ($index->hasFlag('nonclustered')) {
+            $type .= 'NONCLUSTERED ';
+        }
+
+        return $type;
+    }
+
+    /**
      * Extend unique key constraint with required filters
      *
      * @param string $sql
@@ -250,31 +314,55 @@ class MsSqlPlatform extends AbstractPlatform
     public function getAlterTableSQL(TableDiff $diff)
     {
         $queryParts = array();
+        $sql = array();
+        $columnSql = array();
+
         if ($diff->newName !== false) {
-            $queryParts[] = 'RENAME TO ' . $diff->newName;
+            $queryParts[] = "sp_RENAME '" . $diff->name . "', '" . $diff->newName . "'";
         }
 
-        foreach ($diff->addedColumns AS $fieldName => $column) {
+        foreach ($diff->addedColumns as $column) {
+            if ($this->onSchemaAlterTableAddColumn($column, $diff, $columnSql)) {
+                continue;
+            }
+
             $queryParts[] = 'ADD ' . $this->getColumnDeclarationSQL($column->getQuotedName($this), $column->toArray());
         }
 
-        foreach ($diff->removedColumns AS $column) {
+        foreach ($diff->removedColumns as $column) {
+            if ($this->onSchemaAlterTableRemoveColumn($column, $diff, $columnSql)) {
+                continue;
+            }
+
             $queryParts[] = 'DROP COLUMN ' . $column->getQuotedName($this);
         }
 
-        foreach ($diff->changedColumns AS $columnDiff) {
-            /* @var $columnDiff Doctrine\DBAL\Schema\ColumnDiff */
+        foreach ($diff->changedColumns as $columnDiff) {
+            if ($this->onSchemaAlterTableChangeColumn($columnDiff, $diff, $columnSql)) {
+                continue;
+            }
+
+            /* @var $columnDiff \Doctrine\DBAL\Schema\ColumnDiff */
             $column = $columnDiff->column;
-            $queryParts[] = 'CHANGE ' . ($columnDiff->oldColumnName) . ' '
-                    . $this->getColumnDeclarationSQL($column->getQuotedName($this), $column->toArray());
+            $queryParts[] = 'ALTER COLUMN ' .
+                    $this->getColumnDeclarationSQL($column->getQuotedName($this), $column->toArray());
         }
 
-        foreach ($diff->renamedColumns AS $oldColumnName => $column) {
-            $queryParts[] = 'CHANGE ' . $oldColumnName . ' '
-                    . $this->getColumnDeclarationSQL($column->getQuotedName($this), $column->toArray());
+        foreach ($diff->renamedColumns as $oldColumnName => $column) {
+            if ($this->onSchemaAlterTableRenameColumn($oldColumnName, $column, $diff, $columnSql)) {
+                continue;
+            }
+
+            $sql[] = "sp_RENAME '". $diff->name. ".". $oldColumnName . "' , '".$column->getQuotedName($this)."', 'COLUMN'";
+            $queryParts[] = 'ALTER COLUMN ' .
+                    $this->getColumnDeclarationSQL($column->getQuotedName($this), $column->toArray());
         }
 
-        $sql = array();
+        $tableSql = array();
+
+        if ($this->onSchemaAlterTable($diff, $tableSql)) {
+            return array_merge($tableSql, $columnSql);
+        }
 
         foreach ($queryParts as $query) {
             $sql[] = 'ALTER TABLE ' . $diff->name . ' ' . $query;
@@ -282,7 +370,7 @@ class MsSqlPlatform extends AbstractPlatform
 
         $sql = array_merge($sql, $this->_getAlterTableIndexForeignKeySQL($diff));
 
-        return $sql;
+        return array_merge($sql, $tableSql, $columnSql);
     }
 
     /**
@@ -315,7 +403,7 @@ class MsSqlPlatform extends AbstractPlatform
      */
     public function getListTableColumnsSQL($table, $database = null)
     {
-        return 'exec sp_columns @table_name = ' . $table;
+        return "exec sp_columns @table_name = '" . $table . "'";
     }
 
     /**
@@ -420,7 +508,7 @@ class MsSqlPlatform extends AbstractPlatform
     {
         $trimFn = '';
 
-        if (!$char) {
+        if ( ! $char) {
             if ($pos == self::TRIM_LEADING) {
                 $trimFn = 'LTRIM';
             } else if ($pos == self::TRIM_TRAILING) {
@@ -517,6 +605,17 @@ class MsSqlPlatform extends AbstractPlatform
         return 'SMALLINT' . $this->_getCommonIntegerTypeDeclarationSQL($field);
     }
 
+    /**
+     * Decleration for a UNIQUEIDENTIFIER (GUID) field in SQL Server
+     *
+     * @param array $field
+     * @return string
+     */
+    public function getGuidTypeDeclartionSQL(array $field)
+    {
+        return 'UNIQUEIDENTIFIER';
+    }
+
     /** @override */
     protected function getVarcharTypeDeclarationSQLSnippet($length, $fixed)
     {
@@ -548,8 +647,7 @@ class MsSqlPlatform extends AbstractPlatform
      */
     public function getDateTimeTypeDeclarationSQL(array $fieldDeclaration)
     {
-        // 6 - microseconds precision length
-        return 'DATETIME2(6)';
+        return 'DATETIME';
     }
 
     /**
@@ -557,7 +655,7 @@ class MsSqlPlatform extends AbstractPlatform
      */
     public function getDateTypeDeclarationSQL(array $fieldDeclaration)
     {
-        return 'DATE';
+        return 'DATETIME';
     }
 
     /**
@@ -565,7 +663,7 @@ class MsSqlPlatform extends AbstractPlatform
      */
     public function getTimeTypeDeclarationSQL(array $fieldDeclaration)
     {
-        return 'TIME(0)';
+        return 'DATETIME';
     }
 
     /**
@@ -580,27 +678,20 @@ class MsSqlPlatform extends AbstractPlatform
      * Adds an adapter-specific LIMIT clause to the SELECT statement.
      *
      * @param string $query
-     * @param mixed $limit
-     * @param mixed $offset
+     * @param integer $limit
+     * @param integer $offset
      * @link http://lists.bestpractical.com/pipermail/rt-devel/2005-June/007339.html
      * @return string
      */
     protected function doModifyLimitQuery($query, $limit, $offset = null)
     {
         if ($limit > 0) {
-            $count = intval($limit);
-            $offset = intval($offset);
-
-            if ($offset < 0) {
-                throw new DBALException("LIMIT argument offset=$offset is not valid");
-            }
-
             if ($offset == 0) {
-                $query = preg_replace('/^SELECT\s/i', 'SELECT TOP ' . $count . ' ', $query);
+                $query = preg_replace('/^(SELECT\s(DISTINCT\s)?)/i', '\1TOP ' . $limit . ' ', $query);
             } else {
                 $orderby = stristr($query, 'ORDER BY');
 
-                if (!$orderby) {
+                if ( ! $orderby) {
                     $over = 'ORDER BY (SELECT 0)';
                 } else {
                     $over = preg_replace('/\"[^,]*\".\"([^,]*)\"/i', '"inner_tbl"."$1"', $orderby);
@@ -611,13 +702,21 @@ class MsSqlPlatform extends AbstractPlatform
                 $query = preg_replace('/^SELECT\s/', '', $query);
 
                 $start = $offset + 1;
-                $end = $offset + $count;
+                $end = $offset + $limit;
 
-                $query = "SELECT * FROM (SELECT ROW_NUMBER() OVER ($over) AS \"doctrine_rownum\", $query) AS doctrine_tbl WHERE \"doctrine_rownum\" BETWEEN $start AND $end";
+                $query = "SELECT * FROM (SELECT ROW_NUMBER() OVER ($over) AS doctrine_rownum, $query) AS doctrine_tbl WHERE doctrine_rownum BETWEEN $start AND $end";
             }
         }
 
         return $query;
+    }
+
+    /**
+     * @override
+     */
+    public function supportsLimitOffset()
+    {
+        return false;
     }
 
     /**
@@ -660,7 +759,23 @@ class MsSqlPlatform extends AbstractPlatform
      */
     public function getDateTimeFormatString()
     {
-        return 'Y-m-d H:i:s.u';
+        return 'Y-m-d H:i:s.000';
+    }
+
+    /**
+     * @override
+     */
+    public function getDateFormatString()
+    {
+        return 'Y-m-d H:i:s.000';
+    }
+
+    /**
+     * @override
+     */
+    public function getTimeFormatString()
+    {
+        return 'Y-m-d H:i:s.000';
     }
 
     /**
@@ -700,12 +815,9 @@ class MsSqlPlatform extends AbstractPlatform
             'real' => 'float',
             'double' => 'float',
             'double precision' => 'float',
-            'date' => 'date',
             'datetimeoffset' => 'datetimetz',
-            'datetime2' => 'datetime',
             'smalldatetime' => 'datetime',
             'datetime' => 'datetime',
-            'time' => 'time',
             'char' => 'string',
             'varchar' => 'string',
             'text' => 'text',
@@ -713,8 +825,9 @@ class MsSqlPlatform extends AbstractPlatform
             'nvarchar' => 'string',
             'ntext' => 'text',
             'binary' => 'text',
-            'varbinary' => 'text',
+            'varbinary' => 'blob',
             'image' => 'text',
+            'uniqueidentifier' => 'guid',
         );
     }
 
@@ -780,18 +893,23 @@ class MsSqlPlatform extends AbstractPlatform
     }
 
     /**
-     * Quotes a string so that it can be safely used as a table or column name,
-     * even if it is a reserved word of the platform.
-     *
-     * NOTE: Just because you CAN use quoted identifiers doesn't mean
-     * you SHOULD use them.  In general, they end up causing way more
-     * problems than they solve.
-     *
-     * @param string $str           identifier name to be quoted
-     * @return string               quoted identifier string
+     * {@inheritDoc}
      */
-    public function quoteIdentifier($str)
+    public function quoteSingleIdentifier($str)
     {
         return "[" . str_replace("]", "][", $str) . "]";
+    }
+
+    public function getTruncateTableSQL($tableName, $cascade = false)
+    {
+        return 'TRUNCATE TABLE '.$tableName;
+    }
+
+    /**
+     * Gets the SQL Snippet used to declare a BLOB column type.
+     */
+    public function getBlobTypeDeclarationSQL(array $field)
+    {
+        return 'VARBINARY(MAX)';
     }
 }
